PagesServices/pyproject.toml
[tool.poetry]
name = "pagesservices"
version = "0.1.0"
description = ""
authors = ["Alex-Evo-2266 <ghbdtn2266@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.3"
pyyaml = "^6.0.2"
uvicorn = "^0.32.0"
pika = "^1.3.2"
ormar = {extras = ["mysql"], version = "^0.20.2"}
alembic = "^1.16.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
